"Functional" version of the mtsv pre process module

+++ scripts/MTSv_prune.py
@@ -304,7 +304,7 @@ def clip(in_tx,ru_rank, ex_tx, name, min,maximum,fasta_path, pickle_path):
                             out.write(seq)
                         line_count = 0
                         seq = bytearray()
-
+    return os.path.abspath(name)
 # writes a new or updates json config file
 def gen_json(configuration, args):
     if args.update and args.configuration_path:
@@ -472,7 +472,9 @@ def ftp_dl(x):
         fp_path = fp_path[1]
         try:
             outpath = os.path.join(raw_path, os.path.basename(fp_path))
-            # try:
+            if os.path.isfile(outpath.strip(".gz")):
+                continue
+
             file_size = connection.size(fp_path)
             if not os.path.isfile(outpath) or file_size != os.path.getsize(outpath):
                 with open(outpath, "wb") as out_file:
@@ -496,7 +498,7 @@ def ftp_dl(x):
     except:
         pass
 
-def pull(thread_count=1, excluded=set(), path=""):
+def pull(path="",thread_count=1,databases ={"genbank"} ):
     if not path:
         string_date = datetime.datetime.now().strftime("%b-%d-%Y")
     else:
@@ -539,7 +541,7 @@ def pull(thread_count=1, excluded=set(), path=""):
     to_download = []
     level2path = {}
 
-    if "genbank" not in excluded:
+    if "genbank" in databases:
         for fp in connection.nlst(genbank_dir):
             base_fp = os.path.basename(fp)
             for ind,char in enumerate(base_fp):
@@ -573,7 +575,7 @@ def pull(thread_count=1, excluded=set(), path=""):
                 continue
         except:
             pass
-        if line[11].strip().decode() not in excluded:
+        if line[11].strip().decode() in databases:
             try:
                 temp = line[19].split(ftp_path.encode(),1)[1].decode()
                 temp_path = "{0}/{1}_genomic.gbff.gz".format(temp, os.path.basename(temp))
@@ -601,7 +603,7 @@ def pull(thread_count=1, excluded=set(), path=""):
         except:
             pass
 
-        if line[11].strip().decode() not in excluded:
+        if line[11].strip().decode() in databases:
             try:
                 temp = line[19].split(ftp_path.encode(),1)[1].decode()
                 temp_path = "{0}/{1}_genomic.gbff.gz".format(temp, os.path.basename(temp))
@@ -644,7 +646,9 @@ if __name__ =="__main__":
     parser = argparse.ArgumentParser(description="TaxClipper is intended to be used to parse sequences based on NCBI taxid")
 
     group = parser.add_mutually_exclusive_group(required=True)
-    group.add_argument("-oc", "--oneclick", "-oneclick",action='store_true')
+    # group.add_argument("-oc", "--oneclick", "-oneclick",action='store_true')
+    # group.add_argument("-ocdl", "--oneclickdl", "-oneclickdl",action='store_true')
+    # group.add_argument("-ocdc", "--oneclickdc", "-oneclickdc",action='store_true')
 
     group.add_argument("-pl", "--pull", "-pull",action='store_true')
     group.add_argument("-gc", "--generate-config","-generate-config", action='store_true',
